chore: move services into separate folder
services are used across multiple modules, hence, it's easier to keep them all in one place
