+++ webapp/frontend/lib/common/scaffold_with_nav.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:decimal/decimal.dart';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:get_10101/auth/auth_service.dart';
+import 'package:get_10101/services/auth_service.dart';
 import 'package:get_10101/auth/login_screen.dart';
 import 'package:get_10101/common/amount_text.dart';
 import 'package:get_10101/common/balance.dart';
@@ -12,10 +12,10 @@ import 'package:get_10101/common/currency_change_notifier.dart';
 import 'package:get_10101/common/currency_selection_widget.dart';
 import 'package:get_10101/common/model.dart';
 import 'package:get_10101/common/snack_bar.dart';
-import 'package:get_10101/common/version_service.dart';
+import 'package:get_10101/services/version_service.dart';
 import 'package:get_10101/logger/logger.dart';
 import 'package:get_10101/trade/quote_change_notifier.dart';
-import 'package:get_10101/trade/quote_service.dart';
+import 'package:get_10101/services/quote_service.dart';
 import 'package:get_10101/wallet/wallet_change_notifier.dart';
 import 'package:go_router/go_router.dart';
 import 'package:provider/provider.dart';
