+++ webapp/frontend/lib/main.dart
@@ -1,20 +1,20 @@
 import 'package:flutter/material.dart';
-import 'package:get_10101/auth/auth_service.dart';
+import 'package:get_10101/services/auth_service.dart';
 import 'package:get_10101/common/currency_change_notifier.dart';
-import 'package:get_10101/common/version_service.dart';
+import 'package:get_10101/services/version_service.dart';
 import 'package:get_10101/logger/logger.dart';
 import 'package:get_10101/routes.dart';
 import 'package:get_10101/settings/channel_change_notifier.dart';
-import 'package:get_10101/settings/channel_service.dart';
+import 'package:get_10101/services/channel_service.dart';
 import 'package:get_10101/trade/order_change_notifier.dart';
-import 'package:get_10101/trade/order_service.dart';
+import 'package:get_10101/services/order_service.dart';
 import 'package:get_10101/trade/position_change_notifier.dart';
-import 'package:get_10101/trade/position_service.dart';
+import 'package:get_10101/services/position_service.dart';
 import 'package:get_10101/trade/quote_change_notifier.dart';
-import 'package:get_10101/trade/quote_service.dart';
-import 'package:get_10101/settings/settings_service.dart';
+import 'package:get_10101/services/quote_service.dart';
+import 'package:get_10101/services/settings_service.dart';
 import 'package:get_10101/wallet/wallet_change_notifier.dart';
-import 'package:get_10101/wallet/wallet_service.dart';
+import 'package:get_10101/services/wallet_service.dart';
 import 'package:intl/date_symbol_data_local.dart';
 import 'package:intl/intl_browser.dart';
 import 'package:provider/provider.dart';
