+++ webapp/frontend/lib/services/quote_service.dart
