+++ webapp/frontend/lib/services/wallet_service.dart
