+++ webapp/frontend/lib/settings/app_info_screen.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:get_10101/common/color.dart';
 import 'package:get_10101/common/snack_bar.dart';
-import 'package:get_10101/common/version_service.dart';
-import 'package:get_10101/settings/settings_service.dart';
+import 'package:get_10101/services/version_service.dart';
+import 'package:get_10101/services/settings_service.dart';
 import 'package:provider/provider.dart';
 
 class AppInfoScreen extends StatefulWidget {
