+++ webapp/frontend/lib/settings/channel_screen.dart
@@ -4,7 +4,7 @@ import 'package:get_10101/common/channel_state_label.dart';
 import 'package:get_10101/common/snack_bar.dart';
 import 'package:get_10101/common/truncate_text.dart';
 import 'package:get_10101/settings/channel_change_notifier.dart';
-import 'package:get_10101/settings/channel_service.dart';
+import 'package:get_10101/services/channel_service.dart';
 import 'package:get_10101/settings/dlc_channel.dart';
 import 'package:provider/provider.dart';
 import 'package:url_launcher/url_launcher.dart';
